refactor(runner): unify rootfs and snapshot into single image artifact
Replace the separate rootfs and snapshot build/cache/GC lifecycles with
a single unified "image" artifact. This fixes pairing bugs where GC or
cross-machine caching could mismatch a rootfs with an incompatible
snapshot.

Key changes:
- New `images/{image_hash}/` directory replaces `rootfs/{hash}/` and
  `snapshots/{hash}/`
- Single `image_hash` computed from all rootfs + snapshot inputs
- `build.rs` absorbs logic from deleted `rootfs.rs` and `snapshot.rs`
- `ProfileConfig` uses single `image_hash` field
- One flock, one GC unit, one R2 cache key per image
- `runner rootfs` and `runner snapshot` CLI subcommands removed
- CI workflows and Ansible updated to use `image_hash=` output

Closes #8811
