+++ crates/runner/src/cmd/mod.rs
@@ -7,10 +7,8 @@ mod gc;
 mod kill;
 mod local;
 mod nbd;
-mod rootfs;
 mod service;
 mod setup;
-mod snapshot;
 mod start;
 
 pub use benchmark::{BenchmarkArgs, run_benchmark};
@@ -21,8 +19,6 @@ pub use exec::{ExecArgs, run_exec};
 pub use gc::{GcArgs, run_gc};
 pub use kill::{KillArgs, run_kill};
 pub use local::{LocalArgs, run_local};
-pub use rootfs::{RootfsArgs, run_rootfs};
 pub use service::{ServiceArgs, run_service};
 pub use setup::run_setup;
-pub use snapshot::{SnapshotArgs, run_snapshot};
 pub use start::{StartArgs, run_start};
